refactor(platform): split filesystem.rs into filesystem module (Task 5.7)
