+++ crates/katana-platform/src/filesystem/mod.rs
@@ -1,0 +1,6 @@
+//! Filesystem service: reading workspace directories and loading/saving documents.
+
+mod scanner;
+mod service;
+
+pub use service::FilesystemService;
